@@ -3,9 +3,9 @@
   "version": "1.0.0",
   "description": "JavaScript development environment Pluralsight course by Cory House",
   "scripts": {
-    "prestart": "node buildScripts/startMessage.js",
+    "prestart": "babel-node buildScripts/startMessage.js",
     "start": "npm-run-all --parallel open:src",
-    "open:src": "node buildScripts/srcServer.js",
+    "open:src": "babel-node buildScripts/srcServer.js",
     "localtunnel": "lt --port 3000",
     "share": "npm-run-all --parallel open:src localtunnel"
   },
@@ -26,7 +26,7 @@
     "compression": "^1.7.4",
     "cross-env": "3.1.3",
     "css-loader": "^3.0.0",
-    "eslint": "3.8.1",
+    "eslint": ">=4.18.2",
     "eslint-plugin-import": "^2.18.0",
     "eslint-watch": "^5.1.2",
     "express": "^4.17.1",
